# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Template Simbi API",
        "version": "1.0"
    },
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Lista todos os autores cadastrados na API",
                "description": "Lista todos os autores cadastrados.",
                "operationId": "c58eebe54a1725fdda791b1e824149c7",
                "responses": {
                    "200": {
                        "description": "Authors list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Adiciona um novo autor na API",
                "description": "Permite adicionar um autor",
                "operationId": "1f4827c47725ae88eac52beb263f9af4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Lista todos os livros cadastrados na API",
                "description": "Lista todos os livros cadastrados.",
                "operationId": "6b187b4abb5fb71e8505f3ac73313cc8",
                "responses": {
                    "200": {
                        "description": "Books list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Adiciona um novo livro na API",
                "description": "Permite adicionar um novo livro",
                "operationId": "0ed456e5e5823be8d576d3c58547755b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "publisher",
                                    "authorId"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "publisher": {
                                        "type": "string"
                                    },
                                    "authorId": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Author",
                "description": "Author entity",
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "description": "Full name of the author",
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "Timestamp when the author was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the author was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Book": {
                "title": "Book",
                "description": "Book entity",
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "string",
                        "example": "1"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string",
                        "example": "Curso avançado de Java"
                    },
                    "publisher": {
                        "description": "Publisher of the book",
                        "type": "string",
                        "example": "Editora XPTO"
                    },
                    "authorId": {
                        "description": "ID of the author",
                        "type": "string",
                        "example": "1"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "isAvailable": {
                        "description": "Availability status of the book",
                        "type": "boolean",
                        "example": true
                    },
                    "createdAt": {
                        "description": "Timestamp when the book was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the book was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}